% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{add_general_wioa_services}
\alias{add_general_wioa_services}
\title{Add general WIOA services received by the participants}
\usage{
add_general_wioa_services(data)
}
\arguments{
\item{data}{The data frame you want to add the columns to. It should include the columns p1003, p1200, p1300, p1300, p1310, p1315, p1203, p1405, p1409, p1206, and p1412.}
}
\description{
This function adds columns that indicate the high-level services received by the participant. The function creates columns for basic career services, individualized career services, and training services. The function also creates columns for work experience, youth work experience, supportive services, financial literacy, and follow-up services.
}
\examples{

df <- add_general_wioa_services(df)

}
