% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{add_outcome_type_dates}
\alias{add_outcome_type_dates}
\title{Add columns with dates relevant to the WIOA outcome types}
\usage{
add_outcome_type_dates(program_year)
}
\arguments{
\item{program_year}{The program year of the data.}
}
\description{
This function adds columns with dates relevant to the WIOA outcome types. The function adds columns with the program year start and end dates, the quarter 2 start and end dates, the quarter 4 start and end dates, and the quarter 4 roll forward start date.
}
\examples{

df <- add_outcome_type_dates(program_year = 2021)

}
