% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{add_state_information}
\alias{add_state_information}
\title{Add columns with state identification information}
\usage{
add_state_information(
  data,
  source_column,
  state_id_type,
  column_name,
  state_data = state_info
)
}
\arguments{
\item{data}{The data frame you want to add the columns to.}

\item{source_column}{The name of the source column that contains the state identification information.}

\item{state_id_type}{The type of state identification information in the source column. It can be "id", "name", or "fips".}

\item{column_name}{The name of the column you want to add to the data frame.}

\item{state_data}{The data frame that contains the state identification information. The default is the state_info dataframe this is part of the opdR package.}
}
\description{
This function adds colums with state identification information to a data frame. The function identifies the source column as the State ID, State FIPS code, or State full name. It then adds a column with the State Alpha code, Numeric code, or full name to the data frame.
}
\examples{

df <- df |> add_state_information(source_column = "state", state_id_type = "id", column_name = "state_abbreviation")

}
