% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{change_wips_columns}
\alias{change_wips_columns}
\title{Change the Column Names of the Columns from WIPS Warehouse Names to PIRL Names}
\usage{
change_wips_columns(
  data,
  wips_con,
  select_columns = NA,
  version = 2021,
  header = "pirl",
  schema = "pirl"
)
}
\arguments{
\item{data}{The data pulled from the WIPS redshift warehouse}

\item{wips_con}{The WIPS Redshift connection object created to access the tables}

\item{select_columns}{By default this is NA because all columns are selected. This can be changed to a list of columns if you only want to keep certain columns.}

\item{version}{The default is the 2021 version of the PIRL, but older versions are available.}

\item{header}{The default is "pirl" which filters the table to select the names that match the standard PIRL format}

\item{schema}{The defaul is "pirl" which filters the table to select the names that match the standard PIRL format}
}
\description{
This function will automatically change the column names to the standard PIRL names (i.e., pxxx) from the names used in the warehouse.
Right now the function just works for one type of name, but additonal options can be added in the future.
}
\examples{


data <- data \%>\% change_wips_columns(., wips_con = con)

}
