% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{consolidate_reporting_state_column}
\alias{consolidate_reporting_state_column}
\title{Consolidate the reporting state column}
\usage{
consolidate_reporting_state_column(data, program_year)
}
\arguments{
\item{data}{The data frame you want to consolidate the reporting state column in.}

\item{program_year}{The program year of the data.}
}
\description{
This function consolidates the reporting state column in the data frame. It can be needed because the column that identifies the reporting state differs by the program year of the data. The function checks the program year and consolidates the reporting state column to a single column. The function drops the p3000 and p4000 columns if they are present.
}
\examples{

df <- consolidate_reporting_state_column(df, program_year = 2021)

}
