% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{copy_dir_to_SP}
\alias{copy_dir_to_SP}
\title{Copy the whole directory in a local project to the team project folder on SharePoint}
\usage{
copy_dir_to_SP(dir_path, office, project_location = "mirror")
}
\arguments{
\item{dir_path}{The local directory path of the directory you want to copy to SharePoint.}

\item{office}{The office of the SharePoint path you want. This currently set up for OPDR's DASP and DP Teams. To be added email zzETA-DASP@dol.gov}

\item{project_location}{The default is "mirror" which means the directory will be copied to the same location in the SharePoint directory as the local directory. If you want to copy the directory to a different location in the SharePoint directory you can specify the location here.}
}
\description{
This function copies a directory from the users local project to the related SharePoint location
}
\examples{

dir <- here::here("data")
copy_dir_to_SP(dir, office = "DASP")

}
