% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{copy_from_SP}
\alias{copy_from_SP}
\title{Copy a local file in a team SharePoint project to user's local project}
\usage{
copy_from_SP(file_path, office, project_location = "mirror")
}
\arguments{
\item{file_path}{The team SharePoint project path of the file you want to copy to your local. This is typically the file you are about to read in.}

\item{office}{The office of the SharePoint path you want. This currently set up for OPDR's DASP and DP Teams. To be added email zzETA-DASP@dol.gov}

\item{project_location}{The default is "mirror" which means the file will be copied from the same location in the SharePoint directory as the local directory. If you want to copy the file from a different location in the SharePoint directory you can specify the location here.}
}
\description{
This function copies a file from the team project directory on SharePoint to the related project directory on the user's local machine.
}
\examples{

read_path <- here::here("data",  "PY2021.csv")
copy_from_SP(read_path, office = "DASP")
df <- read_csv(read_path)

}
