% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{copy_source_wipr_data}
\alias{copy_source_wipr_data}
\title{Copy a WIPR quarterly data file from the S drive to different location}
\usage{
copy_source_wipr_data(
  program_years,
  destination_dir = here::here("data", "wipr"),
  source_type = "clean",
  py_quarter = 4,
  file_suffix = ""
)
}
\arguments{
\item{program_years}{The WIOA program year files you want. This can be a single year or a vector of years.}

\item{destination_dir}{The path of the directory you want to copy the files to. The default is the data/wipr directory in the project.}

\item{source_type}{The type of WIPR file you want to copy. The default is "clean" which is the SPRA file. The other option is "raw" which is the WIPS file.}

\item{py_quarter}{The quarter of the program year file you want. The default is 4 because it is the file you want to use for annual data.}

\item{file_suffix}{An additional string can be included which adds to the end of the default file name}
}
\description{
This function copies a WIPR quarterly data file from the S drive to a local directory, unzips the file, reads in the csv file, and saves the file as a parquet file.
}
\examples{

copy_source_wipr_data(program_years = 2021)
copy_source_wipr_data(program_years = c(2020, 2021), source_type = "raw", py_quarter = 2)

}
