% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{copy_to_SP}
\alias{copy_to_SP}
\title{Copy a local file in a local project to the team project folder on SharePoint}
\usage{
copy_to_SP(file_path, office, project_location = "mirror")
}
\arguments{
\item{file_path}{The local file path of the file you want to copy to SharePoint. This is typically the file you just wrote to a local directory.}

\item{office}{The office of the SharePoint path you want. This currently set up for OPDR's DASP and DP Teams. To be added email zzETA-DASP@dol.gov}

\item{project_location}{The default is "mirror" which means the file will be copied to the same location in the SharePoint directory as the local directory. If you want to copy the file to a different location in the SharePoint directory you can specify the location here.}
}
\description{
This function copies a file from the users local project to the related SharePoint location
}
\examples{

write_path <- here::here("data",  "PY2021.csv")
write_csv(df, write_path)
copy_to_SP(write_path, office = "DASP")

}
