% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_website_files.R
\name{create_anonymous_files}
\alias{create_anonymous_files}
\title{Creates anonymous named files}
\usage{
create_anonymous_files(files_dir, year, product)
}
\arguments{
\item{files_dir}{The directory where the files are located.}

\item{year}{The program year of the reports/dashboards you want to rename.}

\item{product}{A character string indicating the product type. Options are "workforce_services_dashboards" and "state_assessments".}
}
\description{
This function creates anonymous named files for the DOL website and a spreadsheet with the file/link info.
}
\examples{

create_anonymous_files(files_dir = here::here("reports"), year = 2021, product = "workforce_services_dashboards")

}
