% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{create_wioa_program_columns}
\alias{create_wioa_program_columns}
\title{Create WIOA program columns}
\usage{
create_wioa_program_columns(data, program = "wioa")
}
\arguments{
\item{data}{The data frame you want to add the columns to. It should include the columns p903, p904, p905, and p918.}

\item{program}{The default is "wioa" which will add columns of the 2 core programs. Option"jc" will add a job corps column.}
}
\description{
This function creates columns for each WIOA program (Adult, Dislocated Worker, Youth, and Wagner-Peyser) and indicates if the person is in the program based on the values in the p903, p904, p905, and p918 columns.
}
\examples{

df <- create_wioa_program_columns(df)

}
