% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{format_zip_code}
\alias{format_zip_code}
\title{Format a column with zip codes}
\usage{
format_zip_code(data, zip_column)
}
\arguments{
\item{data}{The data frame that contains the zip code column.}

\item{zip_column}{The name of the zip code column in the data frame.}
}
\description{
This function formats zip code columns to have 5 characters. It pads the zip code with zeros if it has less than 5 characters.
}
\examples{

df <- format_zip_code(df, zip_column = "zip_code")

}
