% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{generate_wioa_outcomes}
\alias{generate_wioa_outcomes}
\title{Generate WIOA outcomes}
\usage{
generate_wioa_outcomes(df, period_start, period_end, msg_restricted = FALSE)
}
\arguments{
\item{df}{The data frame that contains the data you want to generate outcomes for.}

\item{period_start}{The start date of the period you want to generate outcomes for.}

\item{period_end}{The end date of the period you want to generate outcomes for.}

\item{msg_restricted}{A logical value that indicates if the function should generate outcomes for all participants or only for participants that are included per the official rules used to calculate the MSG performance indicator in WIOA. The default is FALSE.}
}
\description{
This function generates WIOA outcomes based on the data provided. The function creates columns that indicate if the participant received a credential, entered employment, retained employment, and received a measurable skill gain. For MST, the function can be used to generate outcomes for all participants in the data or only for participants that are included per the official rules used to calculate the MSG performance indicator in WIOA.
}
\examples{

df <- generate_wioa_outcomes(df, period_start = ymd("2021-07-01"), period_end = ymd("2022-06-30"))

}
