% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{get_main_SP_directory}
\alias{get_main_SP_directory}
\title{Get the main SharePoint directory path}
\usage{
get_main_SP_directory(office)
}
\arguments{
\item{office}{The office of the SharePoint path you want. This currently set up for OPDR's DASP and DP Teams. To be added email zzETA-DASP@dol.gov}
}
\description{
A function to create a string of the path to the chosen office's main SharePoint directory
}
\examples{
DASP_path <- get_main_SP_directory(office = "DASP")
DP_path <- get_main_SP_directory(office = "DP")

}
