% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{list_of_wioa_states}
\alias{list_of_wioa_states}
\title{Create a list of WIOA state abbreviations}
\usage{
list_of_wioa_states(include_territories = FALSE)
}
\arguments{
\item{include_territories}{A logical value that indicates if territories should be included in the list. The default is FALSE.}
}
\description{
This function creates a vector that lists all the state abbreviations (includes DC and PR). Territories can be included as well.
}
\examples{

states <- list_of_wioa_states()
all_eta_areas <- list_of_wioa_states(include_territories = TRUE)

}
