% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RUCA_zipcode}
\alias{RUCA_zipcode}
\title{Rural-Urban Communting Area Codes}
\format{
A data frame
\describe{
\item{ZIP_CODE}{The 5 digit zip code in a character format}
\item{STATE}{The abbreviation of the state or territory}
\item{ZIP_TYPE}{The type of zip code area}
\item{RUCA1}{The primary RUCA codes}
\item{RUCA2}{The secondary RUCA codes}
}
}
\source{
\url{https://www.ers.usda.gov/data-products/rural-urban-commuting-area-codes}
}
\usage{
RUCA_zipcode
}
\description{
A dataset containing codes on rural or urban by zip code.
}
\keyword{datasets}
