% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_info}
\alias{state_info}
\title{State Information}
\format{
A data frame
\describe{
\item{Name}{The full name of the state or territory}
\item{Alpha_code}{The abbreviation of the state or territory}
\item{Status}{The type of area}
\item{Region}{The ETA Regional Office of the area}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_U.S._state_and_territory_abbreviations}
}
\usage{
state_info
}
\description{
A dataset containing naming information for all states and territories.
}
\keyword{datasets}
