% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{connect_to_WIPS}
\alias{connect_to_WIPS}
\title{Connect to WIPS}
\usage{
connect_to_WIPS()
}
\description{
This function creates a connection to the WIPS redshift data and stores that connection to the object name assigned.
}
\examples{

con <- connect_to_WIPS()

}
